requirements.txt
openai==0.24.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
requests==2.28.1
python-dotenv==0.21.0
